Less consing in ivy--re-filter
Avoid repeated cl-remove when matching multiple regexps;
unconditionally make a copy and use cl-delete instead.  This
slightly penalizes the case where no candidates are removed.

Using cl-remove for the first pass and subsequently switching to
cl-delete would not always work, since it's unknown whether and when
cl-remove returns a full or partial copy of its input list.

Yet another option would be to consolidate all regexps AOT, in the
hope that regexp backtracking in C is faster than looping in Lisp,
but that would involve more work both for the programmer and in
terms of consing.  I'm also not certain about the Ivy semantics for
numbered match groups, for matches that are kept.

* ivy.el (ivy--re-filter): Replace potentially repeated cl-remove
with unconditional copy-sequence followed by cl-delete.  This
produces less garbage in pathological cases.