Fix ivy-completion-common-length
Thanks to Kien Nguyen <kien.n.quang@gmail.com> and
Madhu <enometh@meer.net> for suggestions.

Fixes parts of #1361, #1755, #2879, #3051, #3056,
and https://bugs.gnu.org/76440.
Relates to https://bugs.gnu.org/71419.

* ivy.el (ivy--face-list-p): New compatibility shim.
(ivy-completion-common-length): Use it to handle both atom and list
values for the face property, since it varies across Emacs versions.
Use previous-single-property-change instead of iterating by char.
Improve documentation.
(ivy-completion-in-region): Translate a 'not found' result of
ivy-completion-common-length into a first difference at index 0, not
at the end of the string.  That means, interpret a lack of
completions-first-difference as no commonality.  This is more
consistent with both the definition of ivy-completion-common-length
and completion changes in Emacs 30.

* ivy-test.el (ivy-completion-common-length): Extend tests.