Fix some point placement on exact completion exit
When completing ~/foo and ~/foobar is the only completion,
vanilla completion inserts a trailing space on exit to indicate an
exact match, even though completion-try-completion returned a cons.

This is determined by completion--done which calls try-completion on
the completion-try-completion return value i.s.o. the
completion-all-completions candidate that was inserted.

* ivy.el (ivy-completion-in-region-action): Prefer passing
completion-try-completion 'new string' to completion--done when
possible.