Prefer base-size over ivy-completion-common-length
The latter is very problematic: it is specific to a single
candidate, does not necessarily correspond to the string being
completed, complicates bounds calculations, relies on only partially
documented presentation details rather than the usual API, etc.

When completion-all-completions returns a base-size (which in
practice seems to be most if not all the time), it corresponds more
closely to the substring that is being completed and will be
replaced.

This should address most of the remaining issues in
reports #1361, #1755, #2643, #2879, #3051, #3056, and
https://bugs.gnu.org/76440.

* ivy.el (ivy-completion-common-length): Discourage use.
(ivy-completion-in-region): When completion-all-completions
returns a base-size, prefer that over the
completions-first-difference face to determine the completion
boundaries and initial-input.

Fixes #1755.
Fixes #2643.
Fixes #2879.
Closes #3051.
Fixes #3056.