Fix Eglot completion
The recent changes to ivy-completion-in-region fixed some cases but
broke the base-size calculation for Eglot, which doesn't seem to
compute field boundaries.  Either way, the result is that
completion-all-completions does not include a base-size, its
candidates do not bear a completions-first-difference, and
completion-try-completion returns a cons even for no/sole match.

* ivy.el (ivy-completion-in-region): Take lack of base-size as
covering the whole completion string.  This is consistent with
completion-boundaries and other minibuffer.el functions.  Remove
resulting dead code.  If completion-try-completion returns a cons,
try detecting no/sole match from completion-all-completions instead.
This avoids an error when calling ivy-completion-in-region-action on
an empty list of candidates in the no match case, avoids erasing and
reinserting a sole match, and in both cases ensures a message is
printed to the user.
(ivy-completion-common-length): Mark as obsolete since it is no
longer used and shouldn't be relied on.
* ivy-test.el (ivy-completion-common-length): Silence resulting
warnings.

Fixes #3068.
Re: #1361.