Fix kill-ring-save for swiper-isearch
The default M-w binding of ivy-kill-ring-save expects ivy--old-cands
to be a list of strings, but swiper-isearch has long switched to a
list of buffer positions instead.

* ivy.el (ivy-kill-ring-save): Prefer use-region-p over
region-active-p.  Use string-join.  Tweak docstring.
* swiper.el: Set swiper and swiper-isearch actions closer to their
corresponding commands.
(swiper--isearch-kill-ring-save): New command modeled after
ivy-kill-ring-save.
(swiper-isearch-map): Remap ivy-kill-ring-save and kill-ring-save to
it.  By default, the former remap is defensive, and only the latter
is needed.

Fixes #3000.